adds private key const, gets chainId from store, improves test
