+++ src/compositions/useWalletConnect.ts
@@ -2,7 +2,7 @@ import WalletConnectProvider from "@walletconnect/web3-provider";
 import { setState, useState, getState } from "@/stores";
 import useWeb3 from "@/compositions/useWeb3";
 import { provider as Provider } from "web3-core";
-import { DEFAULT_NETWORK_CONFIG } from "@/utils/networkConfig";
+import { DEFAULT_NETWORK_CONFIG } from "@/helpers/config";
 
 let provider: WalletConnectProvider;
 export const WALLET_CONNECT_VERSION = "1.0";
